requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coverage==5.5
Django==2.2.9
django-appconf==1.0.4
django-debug-toolbar==3.2
flake8==3.8.4
flake8-quotes==3.2.0
iniconfig==1.1.1
isort==5.7.0
mccabe==0.6.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==3.8.0
pytz==2021.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
